requirements.txt
mysql-connector-python==8.0.33
requests==2.31.0
Flask==3.0.3
Flask-WTF==1.2.1
python-dotenv==1.0.0